+++ flinkboot-core/src/main/java/io/github/sekelenao/flinkboot/core/internal/startup/StartupEnvironment.java
@@ -60,6 +60,7 @@ public final class StartupEnvironment {
         return ParserFeatures.builder()
             .permitOverride(flag("flinkboot-configuration-override"))
             .listMerging(flag("flinkboot-configuration-list-merging"))
+            .disableValidation(flag("flinkboot-configuration-disable-validation"))
             .validationCapacity(validationCapacity)
             .build();
     }